+++ src/liberator/graph.clj
@@ -10,13 +10,22 @@
 (defn to-graph [[& args]]
   (condp = (first args)
     'defdecision
-    (let [[name then else] (apply extract args)]
-      (format (str "\"%s\" [id = \"%s\"] \n "
-                   "\"%s\" -> \"%s\" [label = \"true\",  id = \"%s\"] \n"
-                   "\"%s\" -> \"%s\" [label = \"false\", id = \"%s\"]\n")
-              name (clean-id name)
-              name then (clean-id (str name "_" then))
-              name else (clean-id (str name "_" else))))
+   (let [[name then else] (apply extract args)
+          internal? (#{"is-options?"
+                       "method-put?"
+                       "method-delete?"
+                       "method-patch?"
+                       "method-post?"
+                       "post-to-existing?"
+                       "post-to-missing?"
+                       "post-to-gone?"
+                       "put-to-existing?"} (str name))]
+      (format (str "\"%s\" [id = \"%s\" %s] \n "
+                   "\"%s\" -> \"%s\" [label = \"true\", id = \"%s\"] \n"
+                   "\"%s\" -> \"%s\" [label=\"false\", id = \"%s\"]\n")
+              name (clean-id name) (if internal? "shape=\"octagon\" style=\"filled\" fillcolor=\"#CCCCCC\"" "")
+              name then (clean-id (str name "_" then)) 
+              name else (clean-id (str name "_" else ))))
     'defaction
     (let [[_ name then] args]
       (format (str "\"%s\"[shape=\"ellipse\" id = \"%s\"];\n"
@@ -49,6 +58,7 @@
 (defn rank-handler-groups [handlers]
   (->> handlers
        (group-by (fn [[name status]] (int (/ status 100))))
+       (remove #(#{4 5} (first %)))
        vals
        (map (fn [sg] (map first sg)))
        (map rank-same)
@@ -63,6 +73,8 @@
         decisions (->> nodes
                       (filter #(= 'defdecision (first %)))
                       (map second))
+        conneg-decisions (filter #(.endsWith (name %) "available?")
+                                   decisions)
         handlers (->> nodes
                       (filter #(= 'defhandler (first %)))
                       (map (fn [[_ name status _]] [name status])))
@@ -71,16 +83,18 @@
                       (map second))]
     {:nodes nodes
      :decisions decisions
+     :conneg-decisions conneg-decisions
      :handlers handlers
      :actions actions}))
 
 (defn generate-graph-dot []
-  (let [{:keys [nodes handlers actions]} (parse-source-definitions)]
+  (let [{:keys [nodes conneg-decisions handlers actions]} (parse-source-definitions)]
     (->> nodes
          (map to-graph)
-         (filter identity)
+         (remove nil?)
          (concat (rank-handler-groups handlers))
          (concat (rank-same (remove #{'initialize-context} actions)))
+         (concat (rank-same (remove #{'initialize-context} conneg-decisions)))
          (apply str)
          (format (str "digraph{\nid=\"trace\"; size=\"1000,1000\"; page=\"1000,1000\";\n\n"
                       "edge[fontname=\"sans-serif\"]\n"